Host resume in separate repo
In the resume repo, I set up an action to compile the LaTeX resume file.
I also set up a github pages in this repo so that the resume cna be
published. Now, this site will always contain the same version of my
resume as on the main branch of the resume repo.

I'll be doing something similar with the math summary PDFs, meaning
eventually there will no longer be and PDF assets.